% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{condition}
\alias{condition}
\title{Condition (density of voles) for event to happen (event root in deSolve parckage) And the events (input of broma in the field)}
\usage{
condition(Vole.Level, Broma.Level)
}
\arguments{
\item{Vole.Level}{A numeric value defining the level of vole for treatment}

\item{Broma.Level}{A numeric value defining the quantity of broma at treatment}
}
\value{
A list of two elements : the trigger moment of the event, and the event
}
\description{
Take a numeric and return the trigger condition of the event
}
