% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{fix_pVar_pars}
\alias{fix_pVar_pars}
\title{fix set of parameters}
\usage{
fix_pVar_pars(params, p)
}
\arguments{
\item{params}{set of parameter as reference}

\item{p}{rate of variation compared to reference}
}
\description{
fix set of parameters
}
