% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ggplot_dual_axis}
\alias{ggplot_dual_axis}
\title{ggplot2 dual axes}
\usage{
ggplot_dual_axis(plot1, plot2, which.axis = "x")
}
\description{
ggplot2 dual axis
}
\details{
see https://gist.github.com/jslefche/e4c0e9f57f0af49fca87
}
\examples{
# Create fake data.frame
data.add.x = data.frame(
  y1 = runif(100, 0, 100),
  x1 = runif(100, 0, 100)
)
# Add second x-axis that scales with first
data.add.x$x2 = (data.add.x$x1 + 50)^0.75

# Create plots
plot1.x = qplot(y = y1, x = x1, data = data.add.x)
plot2.x = qplot(y = y1, x = x2, data = data.add.x)

# Run function
ggplot_dual_axis(plot1.x, plot2.x, "x")

}
