% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R, R/plot.R
\name{graphique_2}
\alias{graphique_2}
\title{Graphic for results of the system}
\usage{
graphique_2(n, times, model.output)

graphique_2(n, times, model.output)
}
\arguments{
\item{n}{A numeric value of the number of simulation}

\item{times}{A numeric vector of the period of integration of the differential equation}

\item{model.output}{The output of the ode integration with deSolve}

\item{n}{A numeric value of the number of simulation}

\item{times}{A numeric vector of the period of integration of the differential equation}

\item{model.output}{The output of the ode integration with deSolve}
}
\value{
A graphic

A graphic
}
\description{
Take the output of the ode integration with deSolve and return graphics

Take the output of the ode integration with deSolve and return graphics
}
