% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solver.R
\name{grid.arrange}
\alias{grid.arrange}
\title{Arrange multiple grobs on a page}
\description{
function from package \link{gridExtra}.
Set up a gtable layout to place multiple grobs on a page.
}
\keyword{internal}
