% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solver.R
\name{init}
\alias{init}
\title{Parameters for the system of differential equations to integrate with deSolve package}
\usage{
init(n.simu, C.soil = 0, C.V = 0, C.M = 0, C.F = 0, D.V = 100,
  D.Vd = 0, D.M = 0.05, D.F = 0.025)
}
\arguments{
\item{n.simu}{number of simulation to take into account}

\item{C.soil}{Concentration broma available in the soil}

\item{C.V}{Concentration broma in individual vole}

\item{C.M}{Concentration broma in individual mustelids}

\item{C.F}{Concentration broma in individual fox}

\item{D.V}{Density of the vole population}

\item{D.Vd}{Density of the dead vole sub-population}

\item{D.M}{Density of the mustelid population}

\item{D.F}{Density of the fox population}
}
\value{
A list of numeric containing the initial condition for the differential equations to integrate with deSolve package
}
\description{
Take all parameters a vector and return a list
}
