% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solver.R
\name{intake.V}
\alias{intake.V}
\title{Intake rate of contaminant for voles}
\usage{
intake.V(C, max.intake.C, D50.intake.C)
}
\arguments{
\item{C}{A numeric value/vector of contaminant concentration available}

\item{max.intake.C}{A numeric value of the maximal intake concentration}

\item{D50.intake}{A numeric value of the concentration in environment at 50% of max intake}
}
\value{
The Intake rate
}
\description{
Take a vector containing
}
