% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_predation.R
\name{Nbr.mortality}
\alias{Nbr.mortality}
\title{Repartition of the mortality in vole and mustelid populations}
\usage{
Nbr.mortality(table_df, params, group.num = 1)
}
\arguments{
\item{table_df}{A data.frame of ode system}

\item{params}{Parameter of the system}

\item{group.num}{Group number to apply the computation}
}
\value{
A list of data.frame
}
\description{
Take the output of the ode integration with deSolve and return a list of data.frame
}
