% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solver.R
\name{ode.broma.tritrophic}
\alias{ode.broma.tritrophic}
\title{The system of differential equations to integrate with deSolve package}
\usage{
ode.broma.tritrophic(Time, State, Pars)
}
\value{
A derivative list to integrate with deSolve package
}
\description{
Take Time, State, Pars as input, return derivative list
}
