% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solver.R
\name{ode}
\alias{ode}
\title{General Solver for Ordinary Differential Equations}
\description{
function from package \link{deSolve}.
Solves a system of ordinary differential equations;
 a wrapper around the implemented ODE solvers
}
\keyword{internal}
