% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_solver.R
\name{pars}
\alias{pars}
\title{Parameters for the system of differential equations to integrate with deSolve package}
\usage{
pars(n = 2, r.V = log(2 * 600)/365, K.V = 600, a.M = 1/30,
  a.Md = 1/30, a.F = 1/10, a.Fd = 1/10, h.M = 1/3.5, h.F = 1/6,
  m.M = 1/(0.8 * 365), d = 1/7, a.FM = 1/10 * 80/300,
  r.F = log(3)/365, K.F = 0.03, e.M = 0.0025, B.V = 80,
  B.M = 300, B.F = 5800, max.intake.C = 6 * 1/5,
  D50.intake.C = 100, LD.50.V = 2, days.to.die.V = 6,
  LD.50.M = 2.1, days.to.die.M = 5, LD.50.F = 0.5,
  days.to.die.F = 5, H.V = 4, H.M = 4, H.F = 2, eta.M = 0.5,
  eta.F = 0.5, k.out.V = 0.4, k.out.M = 0.6, k.out.F = 0.6,
  k.0 = (0.106 + 0.057)/2)
}
\value{
A list of parameters
}
\description{
Take all parameters a vector and return a list
}
