% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R, R/plot.R
\name{pkgLoad}
\alias{pkgLoad}
\title{Check if R package is installed then load library}
\usage{
pkgLoad(x)

pkgLoad(x)
}
\arguments{
\item{x}{A character corresponding to the package to load}

\item{x}{A character corresponding to the package to load}
}
\description{
Take a package x check if installed, install it if necessary and load it: Check if R package is installed then load library

Take a package x check if installed, install it if necessary and load it: Check if R package is installed then load library
}
