% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_predation.R
\name{plotIngest}
\alias{plotIngest}
\title{Plot of the repartion of population ingested}
\usage{
plotIngest(df.Nbr.ingest)
}
\arguments{
\item{df.Nbr.ingest}{A data.frame of ode system}
}
\value{
A graphic
}
\description{
Take the output of the ode integration with deSolve and return a graphic
}
