% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_predation.R
\name{plotMortality}
\alias{plotMortality}
\title{Plot of the repartion of population mortality}
\usage{
plotMortality(df.Nbr.mortality)
}
\arguments{
\item{df.Nbr.ingest}{A data.frame of ode system}
}
\value{
A graphic
}
\description{
Take the output of the ode integration with deSolve and return a graphic
}
