% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{rand_pVar_pars}
\alias{rand_pVar_pars}
\title{random set of parameters}
\usage{
rand_pVar_pars(params, p = 0.1)
}
\arguments{
\item{params}{set of parameter as reference}

\item{p}{rate of variation min max in auniform distribution compared to reference}
}
\description{
random set of parameters
}
