% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{tab_sensitivity}
\alias{tab_sensitivity}
\title{Table of sensitivity for set, min, max values}
\usage{
tab_sensitivity(model_output, params, tab_ref)
}
\arguments{
\item{model_output}{the output of the ode solver}

\item{params}{Parameter of the system}

\item{key}{specific name of the parameters to consider}
}
\value{
A list of data.frame
}
\description{
Table of sensitivity for set, min, max values
}
