% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R, R/plot.R
\name{tableau}
\alias{tableau}
\title{Table to use in graphics}
\usage{
tableau(model.output)

tableau(model.output)
}
\arguments{
\item{model.output}{The output of the ode integration with deSolve}

\item{model.output}{The output of the ode integration with deSolve}
}
\value{
A data.frame

A data.frame
}
\description{
Table to use in graphics

Table to use in graphics
}
